buy access pda is done in client side
