+++ frontend_ts/src/app/layout.tsx
@@ -1,3 +1,4 @@
+"use client"
 import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
@@ -9,10 +10,10 @@ import {
 
 const inter = Inter({ subsets: ['latin'] })
 
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-}
+// export const metadata: Metadata = {
+//   title: 'Create Next App',
+//   description: 'Generated by create next app',
+// }
 
 export default function RootLayout({
   children,
@@ -23,6 +24,8 @@ export default function RootLayout({
     <html lang="en">
       <body className={inter.className}>
         <Wallet>
+          <WalletMultiButton />
+          <WalletDisconnectButton/>
           <h1 className='text-2xl'>market place</h1>
           {children}
         </Wallet>
